final_repo/repo/requirements.txt
streamlit>=1.36.0
faiss-cpu>=1.7.4
sentence-transformers>=2.2.0
pypdf>=3.0.0
requests>=2.28.0
numpy>=1.24.0
pandas>=2.0.0
tabulate>=0.9.0
